+++ lizard_progress/tests/test_models.py
@@ -38,6 +38,8 @@ class OrganizationF(factory.Factory):
     FACTORY_FOR = models.Organization
 
     name = "Test organization"
+    allows_non_predefined_locations = False
+    is_project_owner = True
 
 
 class UserProfileF(factory.Factory):